+++ src/Form/Tool/Service/LtiServiceClientType.php
@@ -98,12 +98,12 @@ class LtiServiceClientType extends AbstractType
                 ]
             )
             ->add(
-                'accept',
+                'media',
                 TextType::class,
                 [
                     'required' => false,
-                    'label' => 'Accept',
-                    'help' => "Request content type to perform the call with"
+                    'label' => 'Media type',
+                    'help' => 'Request accept or content type header to perform the call with'
                 ]
             )
             ->add(